+++ test/common_test/joint_features.cc
@@ -15,22 +15,20 @@
  *
 */
 
-#include <cmath>
 #include <cstddef>
+#include <iostream>
 #include <string>
 #include <gtest/gtest.h>
 
 #include <chrono>
 
 #include <gz/common/Console.hh>
-#include <gz/math/Pose3.hh>
 #include <gz/plugin/Loader.hh>
 
 #include <gz/math/Helpers.hh>
 #include <gz/math/Vector3.hh>
 #include <gz/math/eigen3/Conversions.hh>
 
-#include "gz/physics/detail/FeatureList.hh"
 #include "test/TestLibLoader.hh"
 #include "test/Utils.hh"
 #include "Worlds.hh"
@@ -52,10 +50,7 @@
 #include <gz/physics/sdf/ConstructModel.hh>
 #include <gz/physics/sdf/ConstructWorld.hh>
 
-#include <sdf/Link.hh>
-#include <sdf/Model.hh>
 #include <sdf/Root.hh>
-#include <sdf/World.hh>
 
 template <class T>
 class JointFeaturesTest:
@@ -91,215 +86,6 @@ class JointFeaturesTest:
   public: std::set<std::string> pluginNames;
   public: gz::plugin::Loader loader;
 };
-
-
-struct BasicJointFeatureList : gz::physics::FeatureList<
-    gz::physics::ForwardStep,
-    gz::physics::GetBasicJointProperties,
-    gz::physics::GetBasicJointState,
-    gz::physics::GetJointFromModel,
-    gz::physics::GetModelFromWorld,
-    gz::physics::SetBasicJointState,
-    gz::physics::sdf::ConstructSdfWorld
-> { };
-
-template <class T>
-class BasicJointFeaturesTest : public JointFeaturesTest<T>
-{
-  public: void SetUp() override
-  {
-    JointFeaturesTest<T>::SetUp();
-    const sdf::Errors errors =
-        this->root.Load(common_test::worlds::kSinglePendulumWithBase);
-    ASSERT_TRUE(errors.empty()) << errors.front();
-
-    this->sdfWorld = this->root.WorldByIndex(0);
-    ASSERT_NE(nullptr, this->sdfWorld);
-    this->sdfPendulumModel = this->sdfWorld->ModelByName(this->modelName);
-    ASSERT_NE(nullptr, this->sdfPendulumModel);
-    this->armLink = this->sdfPendulumModel->LinkByName("arm");
-    ASSERT_NE(nullptr, this->armLink);
-
-    auto dur = std::chrono::duration<double>(dt);
-
-    this->input.Get<std::chrono::steady_clock::duration>() =
-      std::chrono::duration_cast<std::chrono::steady_clock::duration>(dur);
-  }
-
-  void InitPluginAndWorld(const std::string &_pluginName)
-  {
-    this->plugin = this->loader.Instantiate(_pluginName);
-
-    this->engine =
-        gz::physics::RequestEngine3d<BasicJointFeatureList>::From(this->plugin);
-    ASSERT_NE(nullptr, engine);
-
-    this->world = this->engine->ConstructWorld(*this->sdfWorld);
-    ASSERT_NE(nullptr, this->world);
-    this->model = this->world->GetModel(this->modelName);
-    ASSERT_NE(nullptr, model);
-    this->joint = this->model->GetJoint(this->jointName);
-    ASSERT_NE(nullptr, this->joint);
-  }
-
-  void Step()
-  {
-    this->world->Step(this->output, this->state, this->input);
-  }
-
-  const std::string modelName{"pendulum_with_base"};
-  const std::string jointName{"upper_joint"};
-  // Hardcode these to test proper SDFormat parsing
-  const double kGravity = 9.81;
-  const double kArmLength = 1.0;
-  const double kRadius = 0.1;
-  const double kMass = 6.0;
-  const double moiCom =
-      1.0 / 12 * kMass * (3 * std::pow(kRadius, 2) + std::pow(kArmLength, 2));
-  // Parallel axis theoerem
-  const double moiPivot = moiCom + kMass * std::pow(kArmLength / 2.0, 2);
-
-  const double dt = 0.001;
-  gz::physics::ForwardStep::Output output;
-  gz::physics::ForwardStep::State state;
-  gz::physics::ForwardStep::Input input;
-
-  sdf::Root root;
-  sdf::World *sdfWorld;
-  sdf::Model *sdfPendulumModel;
-  sdf::Link *armLink;
-  gz::plugin::PluginPtr plugin;
-  gz::physics::Engine3dPtr<BasicJointFeatureList> engine;
-  gz::physics::World3dPtr<BasicJointFeatureList> world;
-  gz::physics::Model3dPtr<BasicJointFeatureList> model;
-  gz::physics::Joint3dPtr<BasicJointFeatureList> joint;
-};
-
-using BasicJointFeaturesTestTypes = ::testing::Types<BasicJointFeatureList>;
-TYPED_TEST_SUITE(BasicJointFeaturesTest, BasicJointFeaturesTestTypes);
-
-TYPED_TEST(BasicJointFeaturesTest, GetSetBasicState)
-{
-  const double kTol = 1e-6;
-  for (const std::string &name : this->pluginNames)
-  {
-    // Some of the basic expectations, such as the initial joint position, are
-    // different in Bullet. Since the implementation is likely to be deprecated,
-    // we will skip it in this test rather than trying to fix the underlying
-    // issue.
-    CHECK_UNSUPPORTED_ENGINE(name, "bullet")
-
-    this->InitPluginAndWorld(name);
-
-    EXPECT_NEAR(0.0, this->joint->GetPosition(0), kTol);
-    EXPECT_NEAR(0.0, this->joint->GetVelocity(0), kTol);
-
-    const double startPos = 0.01;
-    this->joint->SetPosition(0, startPos);
-    EXPECT_NEAR(startPos, this->joint->GetPosition(0), kTol);
-
-    // Expect negative joint velocity after 1 step without joint command
-    // 1 ms time step
-    this->Step();
-    // After step, the position should still be very close to the start position.
-    EXPECT_NEAR(startPos, this->joint->GetPosition(0), 1e-3);
-
-    // Reset the initial conditions.
-    this->joint->SetPosition(0, startPos);
-    this->joint->SetVelocity(0, 0);
-    EXPECT_NEAR(startPos, this->joint->GetPosition(0), kTol);
-    EXPECT_NEAR(0, this->joint->GetVelocity(0), kTol);
-    const int numSteps = 100;
-    for (int i = 0; i < numSteps; ++i)
-    {
-      this->Step();
-    }
-    const double timeElapsed = numSteps * this->dt;
-    EXPECT_LT(this->joint->GetVelocity(0), 0.0);
-    EXPECT_LT(this->joint->GetPosition(0), startPos);
-
-    // Using small angle approximation, the joint position is given by:
-    // θ = A * cos(ω₀ * t) + B * sin (ω₀ * t)
-    // where A = θ₀,
-    // B = 0 since starting angular velocity is 0
-    // ω₀  = √(m * g *l/I)
-    // θ = θ₀ * cos(√(m * g *l/I) * t)
-    // See https://ocw.mit.edu/courses/8-01sc-classical-mechanics-fall-2016/mit8_01scs22_chapter24.pdf
-    const double angFreq = std::sqrt(this->kMass * this->kGravity *
-                                     (this->kArmLength / 2) / this->moiPivot);
-    const double expectedPosition = startPos * std::cos(angFreq * timeElapsed);
-    const double expectedAngVel =
-        -angFreq * startPos * std::sin(angFreq * timeElapsed);
-    // Relax tolerence since we are using small angle approximation.
-    EXPECT_NEAR(expectedPosition, this->joint->GetPosition(0), kTol * 10);
-    EXPECT_NEAR(expectedAngVel, this->joint->GetVelocity(0), kTol);
-  }
-}
-
-TYPED_TEST(BasicJointFeaturesTest, GetSetForceAccel)
-{
-  const double kTol = 1e-6;
-
-  for (const std::string &name : this->pluginNames)
-  {
-    // Some of the basic expectations, such as the initial joint position, are
-    // different in Bullet. Since the implementation is likely to be deprecated,
-    // we will skip it in this test rather than trying to fix the underlying
-    // issue.
-    // We're also skipping bullet-featherstone because the expected joint
-    // position and velocity values do not match what's reported by the physics
-    // engine with a small enough tolerance.
-    CHECK_UNSUPPORTED_ENGINE(name, "bullet")
-    CHECK_UNSUPPORTED_ENGINE(name, "bullet-featherstone")
-
-    this->sdfWorld->SetGravity(gz::math::Vector3d::Zero);
-    this->InitPluginAndWorld(name);
-    const double kForceCmd = 0.5;
-
-    // Setting the force should cause acceleration. We'll check the actual
-    // valuelater.
-    this->joint->SetForce(0, kForceCmd);
-    this->Step();
-
-    // dartsim currently clears all joint forces after a step, so GetForce always return 0.
-    if (this->PhysicsEngineName(name) != "dartsim")
-    {
-      EXPECT_DOUBLE_EQ(kForceCmd, this->joint->GetForce(0));
-    }
-    EXPECT_GT(this->joint->GetAcceleration(0), 0);
-
-    // Stepping without setting force should produce no joint acceleration.
-    this->Step();
-    EXPECT_NEAR(this->joint->GetAcceleration(0), 0, kTol);
-    EXPECT_DOUBLE_EQ(0.0, this->joint->GetForce(0));
-
-    const int numSteps = 100;
-    for (int i = 0; i < numSteps; ++i)
-    {
-      this->joint->SetForce(0, kForceCmd);
-      this->Step();
-    }
-
-    const double timeElapsed = numSteps * this->dt;
-    EXPECT_NEAR(kForceCmd, this->joint->GetAcceleration(0) * this->moiPivot, kTol);
-  }
-}
-
-TYPED_TEST(BasicJointFeaturesTest, GetProperties)
-{
-  const double kTol = 1e-6;
-  for (const std::string &name : this->pluginNames)
-  {
-    CHECK_UNSUPPORTED_ENGINE(name, "bullet")
-
-    this->InitPluginAndWorld(name);
-
-    EXPECT_EQ(2u, this->model->GetJointCount());
-    auto fixedJoint = this->model->GetJoint(0);
-    EXPECT_EQ(0u, fixedJoint->GetDegreesOfFreedom());
-    EXPECT_EQ(1u, this->joint->GetDegreesOfFreedom());
-  }
-}
 
 struct JointFeatureList : gz::physics::FeatureList<
     gz::physics::FindFreeGroupFeature,
